wikipedia-agent/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
cssselect==1.2.0
dataclasses-json==0.5.14
frozenlist==1.4.0
idna==3.4
langchain==0.0.285
langsmith==0.0.35
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.10.6
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
packaging==23.1
pydantic==2.3.0
pydantic_core==2.6.3
python-engineio==4.7.0
python-socketio==5.9.0
PyYAML==6.0.1
readability-lxml==0.8.1
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.20
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
websocket-client==1.6.2
wikipedia==1.4.0
yarl==1.9.2
